# backend/internal/apidocs/openapi.yaml
openapi: 3.1.0
info:
  title: QR Classroom Quiz Platform API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    REST API of the QR Classroom Quiz Platform (single Go binary, see the architecture document).

    **Authentication.** Log in with `POST /api/auth/login`. The server sets an HttpOnly
    `__Host-sid` session cookie (ADR-13); browsers send it automatically. Tokens are never
    returned in a response body.

    **CSRF.** Every state-changing request (POST, PUT, PATCH, DELETE) must send the header
    `X-Requested-With` (any value, e.g. `fetch`). If an `Origin` header is present it must match
    the server's base URL. Swagger UI on this page adds the header for you.

    **Errors** share one shape (`Error`): `code` is stable and machine-readable; validation
    failures (422) list per-field messages in `fields`.

    **Real time.** `GET /ws/attempts/{id}` (student) and `GET /ws/sessions/{id}` (teacher) upgrade
    to WebSockets; messages are JSON objects with a `type` field. See the descriptions on those
    operations.
servers:
  - url: /
security:
  - session: []
...
paths:
...
  /api/admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin, settings]
      summary: Platform default settings
      responses: { '200': { $ref: '#/components/responses/Layer' } }
    put:
      operationId: putAdminSettings
      tags: [admin, settings]
      summary: Set platform default settings
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Overrides' } } } }
      responses: { '200': { $ref: '#/components/responses/Layer' }, '422': { $ref: '#/components/responses/Invalid' } }
...
  /api/teacher/settings:
    get:
      operationId: getTeacherSettings
      tags: [settings]
      summary: My teacher-level defaults and what they resolve to
      responses: { '200': { $ref: '#/components/responses/Layer' } }
    put:
      operationId: putTeacherSettings
      tags: [settings]
      summary: Set my teacher-level defaults
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Overrides' } } } }
      responses: { '200': { $ref: '#/components/responses/Layer' }, '422': { $ref: '#/components/responses/Invalid' } }

  # ---------------- classrooms ----------------
...
  /api/teacher/polls:
    get:
      operationId: getTeacherPolls
      tags: [polls]
      summary: The teacher's polls, newest first
      parameters: [{ $ref: '#/components/parameters/page' }, { $ref: '#/components/parameters/size' }, { $ref: '#/components/parameters/q' }, { $ref: '#/components/parameters/dir' }, { name: sort, in: query, schema: { type: string, enum: [created, title, status] } }, { name: status, in: query, schema: { type: string, enum: [draft, open, closed] } }]
      responses: { '200': { description: 'OK', content: { application/json: { schema: { type: object, properties: { total: { type: integer, description: 'Rows matching, on all pages' }, page: { type: integer }, size: { type: integer }, polls: { type: array, items: { $ref: '#/components/schemas/Poll' } } } } } } } }
    post:
      operationId: postTeacherPolls
      tags: [polls]
      summary: Create a poll (draft). Defaults to anonymous, anyone with the code, self-paced, results after answering.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PollInput' } } } }
      responses: { '201': { description: 'Created', content: { application/json: { schema: { $ref: '#/components/schemas/Poll' } } } }, '422': { $ref: '#/components/responses/Invalid' } }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-sid
      description: HttpOnly session cookie set by `POST /api/auth/login` or `/register`.
  parameters:
...
    page: { name: page, in: query, description: 'Page number, from 1 (PL-FR-01)', schema: { type: integer, minimum: 1, default: 1 } }
    size: { name: size, in: query, description: 'Rows per page', schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
    q: { name: q, in: query, description: 'Search text; matches across all pages', schema: { type: string, maxLength: 200 } }
    dir: { name: dir, in: query, description: 'Sort direction; each sort has its own default', schema: { type: string, enum: [asc, desc] } }
...
  responses:
...
    Invalid: { description: 'Validation failed (`validation_failed`, see `fields`)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    Layer:
      description: Overrides at this level and the resolved effective settings
      content: { application/json: { schema: { type: object, properties: { overrides: { $ref: '#/components/schemas/Overrides' }, effective: { $ref: '#/components/schemas/Effective' } } } } }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: validation_failed }
        message: { type: string }
        fields: { type: object, additionalProperties: { type: string } }
...
    Overrides:
      description: |
        Sparse settings for one level (BA §7). Omitted keys inherit from the level above; each key
        may be set only at the levels listed (the platform level may set any key). 0 means "no limit".
      type: object
      additionalProperties: false
      properties:
        countdown_seconds: { type: integer, minimum: 0, maximum: 3600, description: 'teacher, classroom, quiz, session' }
        question_time_limit_sec: { type: integer, minimum: 0, description: 'quiz, question' }
        quiz_time_limit_sec: { type: integer, minimum: 0, description: 'topic, quiz, session, student' }
        admission_mode: { type: string, enum: [manual, auto], description: 'classroom, session' }
        start_mode: { type: string, enum: [countdown, teacher], description: 'teacher, classroom, quiz, session. countdown: each student counts down from admission; teacher: admitted students wait for the teacher''s Start (D-53)' }
        evaluation_method: { type: string, enum: [key, llm, manual], description: 'teacher, quiz, question' }
        llm_key_id: { type: string, description: 'teacher, quiz, question' }
        llm_model: { type: string, description: 'teacher, quiz, question' }
        feedback_mode: { type: string, enum: [none, predefined, ai, both], description: 'teacher, quiz, question' }
        violation_policy: { type: string, enum: [invalidate, warn_then_invalidate, log_only], description: 'classroom, quiz, session' }
        allowed_warnings: { type: integer, minimum: 0, description: 'classroom, quiz, session' }
        blur_grace_ms: { type: integer, minimum: 0, description: 'classroom, quiz, session' }
        disconnect_grace_sec: { type: integer, minimum: 0, description: 'classroom, quiz, session' }
        question_order: { type: string, enum: [fixed, shuffled], description: 'quiz, session' }
        option_order: { type: string, enum: [fixed, shuffled], description: 'quiz, question' }
        one_way_navigation: { type: boolean, description: 'quiz' }
        results_visibility: { type: string, enum: [private, public], description: 'quiz, session' }
        results_view: { type: string, enum: [individual, question_pct, pass_rate], description: 'quiz, session' }
        results_release: { type: string, enum: [immediate, on_session_end, manual], description: 'quiz, session' }
        results_show_answers: { type: boolean, description: 'quiz, session' }
        results_show_correct: { type: boolean, description: 'quiz, session' }
        results_show_feedback: { type: boolean, description: 'quiz, session' }
        pass_mark_pct: { type: integer, minimum: 0, maximum: 100, description: 'classroom, quiz' }
        student_id_required: { type: boolean, description: 'classroom' }
        enrolment_approval: { type: boolean, description: 'classroom' }
        auto_enrol_on_join: { type: boolean, description: 'classroom' }
        team_mode: { type: string, enum: ['off', manual, random, categories, self], description: 'quiz, session (D-44)' }
        team_acceptance: { type: string, enum: [all, first, captain, best], description: 'quiz, session' }
        team_calc: { type: string, enum: [sum, average, max, min], description: 'quiz, session; team_acceptance=all' }
    Effective:
      description: Fully resolved settings; every key from `Overrides` has a value.
      type: object
      additionalProperties: true
...
    Choice: { type: object, properties: { id: { type: string, example: o1 }, text: { type: string } } }
...
    PollSettings:
      type: object
      properties:
        identity: { type: string, enum: [anonymous, identified, optional], description: 'anonymous: nobody is identified (no account is stored, even when logged in); identified: login required; optional: each participant chooses' }
        audience: { type: string, enum: [anyone, classroom], description: 'classroom requires identified and classroom_id' }
        pacing: { type: string, enum: [self, presenter] }
        show_results: { type: string, enum: [live, after_answer, presenter, never], description: 'What participants see. after_answer is applied by each browser; never is enforced by the server.' }
        allow_edit: { type: boolean }
        scoring: { type: boolean, description: 'Questions with an answer key award points (D-42)' }
        speed_bonus: { type: boolean, description: 'Timed presenter-led questions: a right answer earns 50-100% of its points depending on speed' }
        leaderboard: { type: string, enum: ['off', presenter, everyone], default: presenter }
        show_answers: { type: string, enum: [never, after_answer, presenter, after_close], description: 'When participants see the correct answer. presenter needs presenter pacing.' }
        names: { type: string, enum: [nickname, name], description: 'Leaderboard names; name (real names) is not allowed in anonymous polls' }
        groups: { type: string, enum: ['off', manual, random, categories, self], description: 'How groups form (D-43). categories needs a classroom and participants who log in.' }
        group_acceptance: { type: string, enum: [all, first, captain, best], description: "Which answers count for the group: every member's, the group's first, the captain's, or the best" }
        group_calc: { type: string, enum: [sum, average, max, min], description: 'acceptance=all: how members'' scores combine. Average and lowest count silent members as 0.' }
    PollInput:
      type: object
      properties:
        title: { type: string }
        classroom_id: { type: [string, 'null'], format: uuid }
        settings: { $ref: '#/components/schemas/PollSettings' }
    Poll:
      allOf:
        - $ref: '#/components/schemas/PollSettings'
        - type: object
          properties:
            id: { type: string, format: uuid }
            title: { type: string }
            classroom_id: { type: [string, 'null'], format: uuid }
            join_code: { type: string, example: PK7Q2X }
            join_url: { type: string }
            status: { type: string, enum: [draft, open, closed] }
            current_index: { type: integer }
            revealed: { type: boolean }
            answers_revealed: { type: boolean }
            question_started_at: { type: [string, 'null'], format: date-time }
            participants: { type: integer }
            questions: { type: array, items: { $ref: '#/components/schemas/PollQuestion' } }
    PollQuestionInput:
      type: object
      required: [type, text]
      properties:
        type: { type: string, enum: [SINGLE, MULTI, MATCH, BLANK_OPT, BLANK_TEXT, DRAG, ESSAY, SHORT_TEXT, WORD_CLOUD, NUMBER, DATE, TIME, RATING, SLIDER, LIKERT, MATRIX, FILE, AUDIO, VIDEO, CODE] }
        text: { type: string, description: 'Blanks as [[1]], [[2]] for BLANK_*' }
        required: { type: boolean }
        body: { $ref: '#/components/schemas/PollBody' }
        key: { $ref: '#/components/schemas/PollKey' }
        points: { type: integer, minimum: 0, maximum: 10000, default: 100 }
        time_limit_sec: { type: [integer, 'null'], minimum: 5, maximum: 3600, description: 'Presenter pacing: answers close after this' }
    PollKey:
      type: object
      description: 'Answer key of a scorable question (SINGLE, MULTI, MATCH, BLANK_OPT, BLANK_TEXT, DRAG, SHORT_TEXT, NUMBER, SLIDER, DATE, TIME). Teachers only; participants receive it once revealed.'
      properties:
        correct: { type: array, items: { type: string }, description: 'SINGLE, MULTI: option ids' }
        pairs: { type: object, additionalProperties: { type: string }, description: 'MATCH left id to right id; DRAG item id to box id' }
        blanks: { type: object, additionalProperties: { type: array, items: { type: string } }, description: 'BLANK_OPT option ids or BLANK_TEXT accepted words' }
        order: { type: array, items: { type: string }, description: 'DRAG ranking' }
        accepted: { type: array, items: { type: string }, description: 'SHORT_TEXT answers; DATE or TIME values' }
        value: { type: number, description: 'NUMBER, SLIDER' }
        tolerance: { type: number }
        case_sensitive: { type: boolean }
...
    PollQuestion:
      allOf:
        - $ref: '#/components/schemas/PollQuestionInput'
        - type: object
          properties: { id: { type: string, format: uuid }, poll_id: { type: string, format: uuid }, position: { type: integer } }
    PollBody:
      type: object
      description: Each type uses some fields; ids (o1, l1, r1, z1, row1, col1) are assigned when missing.
      properties:
        format: { type: string, enum: ['', markdown] }
        options: { type: array, items: { $ref: '#/components/schemas/Choice' }, description: 'SINGLE, MULTI, BLANK_OPT, DRAG' }
        display: { type: string, enum: [radio, dropdown], description: SINGLE }
        max_choices: { type: integer, description: 'MULTI (0 = any)' }
        left: { type: array, items: { $ref: '#/components/schemas/Choice' } }
        right: { type: array, items: { $ref: '#/components/schemas/Choice' } }
        zones: { type: array, items: { $ref: '#/components/schemas/Choice' }, description: 'DRAG into boxes; empty = ranking' }
        rows: { type: array, items: { $ref: '#/components/schemas/Choice' }, description: 'MATRIX rows, LIKERT statements' }
        columns: { type: array, items: { $ref: '#/components/schemas/Choice' } }
        mode: { type: string, enum: [single, multi, text], description: MATRIX }
        scale: { type: array, items: { type: string }, description: 'LIKERT: 5 or 7 labels' }
        min: { type: number }
        max: { type: number }
        step: { type: number }
        min_label: { type: string }
        max_label: { type: string }
        unit: { type: string }
        from: { type: string, description: 'DATE YYYY-MM-DD or TIME HH:MM' }
        to: { type: string }
        points: { type: integer, description: 'RATING: 3-10 stars' }
        max_length: { type: integer }
        max_words: { type: integer }
        max_entries: { type: integer, description: 'WORD_CLOUD: 1-5' }
        language: { type: string, description: CODE }
        accept: { type: array, items: { type: string, enum: [image, pdf, document, any] } }
        max_mb: { type: integer, description: 'FILE: 1-5' }
        max_seconds: { type: integer, description: 'AUDIO 5-60, VIDEO 5-30' }
...